Added CSV-Reader and characters.csv for easier data entry
There's a lot of characters in legend of Korra. Having a CSV file that I
can continue to edit and other people could potentially contribute to in
the future makes database work much more automated.

When the app starts, it loads up characters.csv, and automatically
updates the database for new or outdated records.