Change data structure of characters in DB
I need to be able to sort characters in the future based off a field.
The easiest way to do this would be to add an ID to the character in the
database, and auto-ID the documents upon creation. Because I'm
auto-ID'ing the documents in the DB, the way I structure and query the
DB will have to change.